+++ np.Templating/CHANGELOG.md
@@ -16,6 +16,8 @@ See Plugin [Documentation](https://noteplan.co/templates/docs) for details on av
 - Improve templating error handling/making suggestions for how to fix on JS code execution errors
 - Add detection/messaging of template function calls called without parentheses
 - Add ability to pass newNoteTitle argument to `templateNew` command and JSON vars for Shortcuts support
+- Added `incrementalRender` setting to allow for turning off incremental render debugging of templates when they fail to render
+- Added `editSettings` command to allow for mobile editing of plugin settings
 
 ## [1.12.0] 2025-03-09 @dwertheimer